Implemented the post system, which supports titles, descriptions, and optional images. 04-05-2026

+++ app/Database/Migrations/2026-04-30-195216_Users.php
@@ -8,39 +8,37 @@ class Users extends Migration
 {
     public function up()
     {
-        //
         $this->forge->addField([
-            'id_user'=>[
-                'type'=>'bigint',
-                'constraint'=>20,
-                'unsigned'=>true,
-                'auto_increment'=>true,
+            'id_user' => [
+                'type' => 'int',
+                'constraint' => 20,
+                'unsigned' => true,
+                'auto_increment' => true,
             ],
-            'user'=>[
-                'type'=>'varchar',
-                'constraint'=>'255',
+            'user' => [
+                'type' => 'varchar',
+                'constraint' => '255',
             ],
-            'email'=>[
-                'type'=>'varchar',
-                'constraint'=>'255',
+            'email' => [
+                'type' => 'varchar',
+                'constraint' => '255',
             ],
-            'password'=>[
-                'type'=>'varchar',
-                'constraint'=>'255',
+            'password' => [
+                'type' => 'varchar',
+                'constraint' => '255',
             ],
-            'type'=>[
-                'type'=>'varchar',
-                'constraint'=>'255',
+            'type' => [
+                'type' => 'varchar',
+                'constraint' => '255',
             ],
         ]);
 
-        $this->forge->addkey('id_user',true);
+        $this->forge->addkey('id_user', true);
         $this->forge->createTable('users');
     }
 
     public function down()
     {
-        //
         $this->forge->dropTable('users');
     }
 }
